Add a script for sorting `extensions.toml` and `.gitmodules`
