+++ package.json
@@ -8,7 +8,8 @@
     "build": "tsc -p .",
     "test": "vitest run",
     "test:watch": "vitest",
-    "package-extensions": "node src/package-extensions.js"
+    "package-extensions": "node src/package-extensions.js",
+    "sort-extensions": "node src/sort-extensions.js"
   },
   "dependencies": {
     "@aws-sdk/client-s3": "3.454.0",
