+++ src/sort-extensions.js
@@ -1,0 +1,5 @@
+import { sortExtensionsToml } from "./lib/extensions-toml.js";
+import { sortGitmodules } from "./lib/git.js";
+
+await sortExtensionsToml("extensions.toml");
+await sortGitmodules(".gitmodules");
